Add info from the network commit message to its doc, split network and its kernel (network is long)
